Project 2 final edit.
I finished all the scripts and edited the flaws that were remaining, this version has descriptive comments written next to the piece of code to explain what the code does and the function it has on the script. Battlesystem script was tweaked and i added a new piece of code to make the if and else statements look shorter and cleaner.